Added footer.js to the components file and extracted into its own file, also changed the App.js file to reflect this. Tested and still working.
