+++ src/components/Footer.js
@@ -1,0 +1,18 @@
+// eslint-disable-next-line
+const Footer = () =>
+{
+  const footerStyle = {
+    color: 'green',
+    fontStyle: 'italic',
+    fontSize: 16
+  }
+
+  return(
+    <div style={footerStyle}>
+      <br />
+      <em>Note App, Jason Kurtzman</em>
+    </div>
+  );
+}
+
+export default Footer;
